+++ src/main/java/ladder/domain/ladder/Items.java
@@ -1,4 +1,6 @@
-package ladder.domain;
+package ladder.domain.ladder;
+
+import ladder.domain.Position;
 
 import java.util.ArrayList;
 import java.util.List;
