+++ src/main/java/ladder/domain/ladder/Ladder.java
@@ -1,4 +1,7 @@
-package ladder.domain;
+package ladder.domain.ladder;
+
+import ladder.domain.player.Player;
+import ladder.domain.Position;
 
 import java.util.ArrayList;
 import java.util.Collections;
