+++ src/main/java/ladder/domain/ladder/LineStrategy.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.ladder;
 
 import java.util.List;
