+++ src/main/java/ladder/domain/ladder/NoncontinuousRandomLineStrategy.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.ladder;
 
 import java.util.ArrayList;
 import java.util.List;
