+++ src/main/java/ladder/domain/ladder/Step.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.ladder;
 
 public enum Step {
     EXIST(true),
