+++ src/test/java/ladder/domain/LadderGameTest.java
@@ -1,5 +1,6 @@
 package ladder.domain;
 
+import ladder.domain.ladder.NoncontinuousRandomLineStrategy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
