+++ src/test/java/ladder/domain/NoncontinuousRandomLineStrategyTest.java
@@ -1,5 +1,7 @@
 package ladder.domain;
 
+import ladder.domain.ladder.NoncontinuousRandomLineStrategy;
+import ladder.domain.ladder.Step;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
